Add Typescript to Devices page
Add types for the Devices page and its related components
